+++ hadoop-ozone/recon/src/main/java/org/apache/hadoop/ozone/recon/api/types/EntityType.java
@@ -35,218 +35,8 @@ import java.util.Set;
  */
 public enum EntityType {
   ROOT {
-    @Override
-    public NamespaceSummaryResponse getSummaryResponse(String[] names, EntityUtils entityUtils) throws IOException {
-      NamespaceSummaryResponse namespaceSummaryResponse = new NamespaceSummaryResponse(EntityType.ROOT);
-      List<OmVolumeArgs> volumes = entityUtils.listVolumes();
-      namespaceSummaryResponse.setNumVolume(volumes.size());
-      List<OmBucketInfo> allBuckets = entityUtils.listBucketsUnderVolume(null);
-      namespaceSummaryResponse.setNumBucket(allBuckets.size());
-      int totalNumDir = 0;
-      long totalNumKey = 0L;
-      for (OmBucketInfo bucket : allBuckets) {
-        long bucketObjectId = bucket.getObjectID();
-        totalNumDir += entityUtils.getTotalDirCount(bucketObjectId);
-        totalNumKey += entityUtils.getTotalKeyCount(bucketObjectId);
-      }
-
-      namespaceSummaryResponse.setNumTotalDir(totalNumDir);
-      namespaceSummaryResponse.setNumTotalKey(totalNumKey);
-
-      return namespaceSummaryResponse;
-    }
-
-    @Override
-    public DUResponse getDuResponse(String path, String[] names, boolean listFile,
-                                    boolean withReplica, EntityUtils entityUtils) throws IOException {
-      DUResponse duResponse = new DUResponse();
-      ReconOMMetadataManager omMetadataManager = entityUtils.getOmMetadataManager();
-      List<OmVolumeArgs> volumes = entityUtils.listVolumes();
-      duResponse.setCount(volumes.size());
-
-      List<DUResponse.DiskUsage> volumeDuData = new ArrayList<>();
-      long totalDataSize = 0L;
-      long totalDataSizeWithReplica = 0L;
-      for (OmVolumeArgs volume: volumes) {
-        String volumeName = volume.getVolume();
-        String subpath = omMetadataManager.getVolumeKey(volumeName);
-        DUResponse.DiskUsage diskUsage = new DUResponse.DiskUsage();
-        long dataSize = 0;
-        diskUsage.setSubpath(subpath);
-        // iterate all buckets per volume to get total data size
-        for (OmBucketInfo bucket: entityUtils.listBucketsUnderVolume(volumeName)) {
-          long bucketObjectID = bucket.getObjectID();
-          dataSize += entityUtils.getTotalSize(bucketObjectID);
-        }
-        totalDataSize += dataSize;
-
-        // count replicas
-        // TODO: to be dropped or optimized in the future
-        if (withReplica) {
-          long volumeDU = entityUtils.calculateDUForVolume(volumeName);
-          totalDataSizeWithReplica += volumeDU;
-          diskUsage.setSizeWithReplica(volumeDU);
-        }
-        diskUsage.setSize(dataSize);
-        volumeDuData.add(diskUsage);
-      }
-      if (withReplica) {
-        duResponse.setSizeWithReplica(totalDataSizeWithReplica);
-      }
-      duResponse.setSize(totalDataSize);
-      duResponse.setDuData(volumeDuData);
-
-      return duResponse;
-    }
-
-    @Override
-    public QuotaUsageResponse getQuotaResponse(String[] names, EntityUtils entityUtils) throws IOException {
-      QuotaUsageResponse quotaUsageResponse = new QuotaUsageResponse();
-      List<OmVolumeArgs> volumes = entityUtils.listVolumes();
-      List<OmBucketInfo> buckets = entityUtils.listBucketsUnderVolume(null);
-      long quotaInBytes = 0L;
-      long quotaUsedInBytes = 0L;
-
-      for (OmVolumeArgs volume: volumes) {
-        final long quota = volume.getQuotaInBytes();
-        assert (quota >= -1L);
-        if (quota == -1L) {
-          // If one volume has unlimited quota, the "root" quota is unlimited.
-          quotaInBytes = -1L;
-          break;
-        }
-        quotaInBytes += quota;
-      }
-      for (OmBucketInfo bucket: buckets) {
-        long bucketObjectId = bucket.getObjectID();
-        quotaUsedInBytes += entityUtils.getTotalSize(bucketObjectId);
-      }
-
-      quotaUsageResponse.setQuota(quotaInBytes);
-      quotaUsageResponse.setQuotaUsed(quotaUsedInBytes);
-      return quotaUsageResponse;
-    }
-
-    @Override
-    public FileSizeDistributionResponse getDistResponse(String[] names, EntityUtils entityUtils) throws IOException {
-      FileSizeDistributionResponse distResponse =
-              new FileSizeDistributionResponse();
-      List<OmBucketInfo> allBuckets = entityUtils.listBucketsUnderVolume(null);
-      int[] fileSizeDist = new int[ReconConstants.NUM_OF_BINS];
-
-      // accumulate file size distribution arrays from all buckets
-      for (OmBucketInfo bucket : allBuckets) {
-        long bucketObjectId = bucket.getObjectID();
-        int[] bucketFileSizeDist = entityUtils.getTotalFileSizeDist(bucketObjectId);
-        // add on each bin
-        for (int i = 0; i < ReconConstants.NUM_OF_BINS; ++i) {
-          fileSizeDist[i] += bucketFileSizeDist[i];
-        }
-      }
-      distResponse.setFileSizeDist(fileSizeDist);
-      return distResponse;
-    }
   },
   VOLUME {
-    @Override
-    public NamespaceSummaryResponse getSummaryResponse(String[] names, EntityUtils entityUtils) throws IOException {
-      NamespaceSummaryResponse namespaceSummaryResponse =
-              new NamespaceSummaryResponse(EntityType.VOLUME);
-      List<OmBucketInfo> buckets = entityUtils.listBucketsUnderVolume(names[0]);
-      namespaceSummaryResponse.setNumBucket(buckets.size());
-      int totalDir = 0;
-      long totalKey = 0L;
-
-      // iterate all buckets to collect the total object count.
-      for (OmBucketInfo bucket : buckets) {
-        long bucketObjectId = bucket.getObjectID();
-        totalDir += entityUtils.getTotalDirCount(bucketObjectId);
-        totalKey += entityUtils.getTotalKeyCount(bucketObjectId);
-      }
-
-      namespaceSummaryResponse.setNumTotalDir(totalDir);
-      namespaceSummaryResponse.setNumTotalKey(totalKey);
-
-      return namespaceSummaryResponse;
-    }
-
-    @Override
-    public DUResponse getDuResponse(String path, String[] names, boolean listFile,
-                                    boolean withReplica, EntityUtils entityUtils) throws IOException {
-      DUResponse duResponse = new DUResponse();
-      ReconOMMetadataManager omMetadataManager = entityUtils.getOmMetadataManager();
-      String volName = names[0];
-      List<OmBucketInfo> buckets = entityUtils.listBucketsUnderVolume(volName);
-      duResponse.setCount(buckets.size());
-
-      // List of DiskUsage data for all buckets
-      List<DUResponse.DiskUsage> bucketDuData = new ArrayList<>();
-      long volDataSize = 0L;
-      long volDataSizeWithReplica = 0L;
-      for (OmBucketInfo bucket: buckets) {
-        String bucketName = bucket.getBucketName();
-        long bucketObjectID = bucket.getObjectID();
-        String subpath = omMetadataManager.getBucketKey(volName, bucketName);
-        DUResponse.DiskUsage diskUsage = new DUResponse.DiskUsage();
-        diskUsage.setSubpath(subpath);
-        long dataSize = entityUtils.getTotalSize(bucketObjectID);
-        volDataSize += dataSize;
-        if (withReplica) {
-          long bucketDU = entityUtils.calculateDUUnderObject(bucketObjectID);
-          diskUsage.setSizeWithReplica(bucketDU);
-          volDataSizeWithReplica += bucketDU;
-        }
-        diskUsage.setSize(dataSize);
-        bucketDuData.add(diskUsage);
-      }
-      if (withReplica) {
-        duResponse.setSizeWithReplica(volDataSizeWithReplica);
-      }
-      duResponse.setSize(volDataSize);
-      duResponse.setDuData(bucketDuData);
-      return duResponse;
-    }
-
-    @Override
-    public QuotaUsageResponse getQuotaResponse(String[] names, EntityUtils entityUtils) throws IOException {
-      QuotaUsageResponse quotaUsageResponse = new QuotaUsageResponse();
-      ReconOMMetadataManager omMetadataManager = entityUtils.getOmMetadataManager();
-      List<OmBucketInfo> buckets = entityUtils.listBucketsUnderVolume(names[0]);
-      String volKey = omMetadataManager.getVolumeKey(names[0]);
-      OmVolumeArgs volumeArgs =
-              omMetadataManager.getVolumeTable().getSkipCache(volKey);
-      long quotaInBytes = volumeArgs.getQuotaInBytes();
-      long quotaUsedInBytes = 0L;
-
-      // Get the total data size used by all buckets
-      for (OmBucketInfo bucketInfo: buckets) {
-        long bucketObjectId = bucketInfo.getObjectID();
-        quotaUsedInBytes += entityUtils.getTotalSize(bucketObjectId);
-      }
-      quotaUsageResponse.setQuota(quotaInBytes);
-      quotaUsageResponse.setQuotaUsed(quotaUsedInBytes);
-      return quotaUsageResponse;
-    }
-
-    @Override
-    public FileSizeDistributionResponse getDistResponse(String[] names, EntityUtils entityUtils) throws IOException {
-      FileSizeDistributionResponse distResponse =
-              new FileSizeDistributionResponse();
-      List<OmBucketInfo> buckets = entityUtils.listBucketsUnderVolume(names[0]);
-      int[] volumeFileSizeDist = new int[ReconConstants.NUM_OF_BINS];
-
-      // accumulate file size distribution arrays from all buckets under volume
-      for (OmBucketInfo bucket : buckets) {
-        long bucketObjectId = bucket.getObjectID();
-        int[] bucketFileSizeDist = entityUtils.getTotalFileSizeDist(bucketObjectId);
-        // add on each bin
-        for (int i = 0; i < ReconConstants.NUM_OF_BINS; ++i) {
-          volumeFileSizeDist[i] += bucketFileSizeDist[i];
-        }
-      }
-      distResponse.setFileSizeDist(volumeFileSizeDist);
-      return distResponse;
-    }
   },
   BUCKET {
     @Override
